+++ tutorial/main.py
@@ -20,6 +20,9 @@ from src.data_handling.TSPETLFactory import TSPETLFactory  # noqa
 from src.data_handling.schemas import schemas
 from src.templates.algorithm import algorithm_templates
 from src.templates.kpi import kpi_templates
+from src.pages.page_scenarios import TSPScenarioPage
+from src.pages.page_compare import TSPComparePage
+from src.pages.page_overview import TSPOverviewPage
 
 
 def configure_styling() -> StylingConfigurator:
@@ -74,6 +77,9 @@ def main() -> None:
         has_persistent_state=True,
         styling_config=configure_styling(),
         title="Algomancy tutorial dashboard",
+        scenario_page=TSPScenarioPage(),
+        compare_page=TSPComparePage(),
+        overview_page=TSPOverviewPage()
     )
 
     # Build the app with AppConfiguration object directly